+++ tools/integration_tests/implicit_dir/implicit_dir_test.go
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Provide tests when implicit directory present and mounted bucket without --implicit-dir flag.
-package without_implicit_dir_flag_test
+// Provide tests when implicit directory present and mounted bucket with --implicit-dir flag.
+package implicit_dir_test
 
 import (
 	"testing"
 
-	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/implicitdir"
+	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup/implicit_and_explicit_dir_setup"
 )
 
 func TestMain(m *testing.M) {
-	flags := [][]string{{"--enable-storage-client-library=true"}, {"--enable-storage-client-library=false"}}
+	flags := [][]string{{"--implicit-dirs"}, {"--enable-storage-client-library=false", "--implicit-dirs"}}
 
-	implicitdir.RunTestsForImplicitDir(flags, m)
+	implicit_and_explicit_dir_setup.RunTestsForImplicitDir(flags, m)
 }
