+++ tools/integration_tests/util/setup/implicit_and_explicit_dir_setup/testdata/create_objects.sh
